Explicitly raise an error when server returns http 403 and dont continue
Previously if the server returned a 403, eventsource would begin a retry
loop to connect to the server over and over. This loop will start before
a client can attach to the 'error' event because eventsource connects
right away in the constructor. This prevents client code from closing
the eventsource connection. This commit addresses this issue and fails
fast instead of allowing eventsource to enter a retry loop when a server
returns a 403.