Remove unneeded isPlainObject function
It's the function caller's responsibility to pass the expected argument type. If they pass something "truthy" that is not an Object object, then the behavior of the function is undefined. We shouldn't work around such mistakes by the caller (i.e. what did they mean to pass for the argument?) and why should we care if the Object object is of the "plain" variety.

This pull request is just a half step in the right direction. A better solution would be to throw an exception on receiving incorrect arguments.